feat: update to algorithm table works
